requirements.txt
aiohttp==3.7.4.post0
aioice==0.7.6
aiortc==1.3.1
altair==4.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.6.0
av==9.1.1
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
discord.py==1.7.3
dnspython==2.2.1
docopt==0.6.2
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
fastapi==0.75.2
Flask==2.1.1
gitdb==4.0.9
GitPython==3.1.27
google-crc32c==1.3.0
h11==0.13.0
httptools==0.4.0
idna==3.3
imageio==2.21.1
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.11.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
kiwisolver==1.3.2
MarkupSafe==2.1.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==2.8.5
notebook==6.4.10
numpy==1.21.3
opencv-python==4.5.4.58
orjson==3.6.8
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==8.4.0
pipreqs==0.4.11
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.20.0
psutil==5.9.0
pure-eval==0.2.2
pyarrow==7.0.0
pycodestyle==2.8.0
pycparser==2.20
pydantic==1.9.0
pydeck==0.7.1
pyee==9.0.4
Pygments==2.11.2
pylibsrtp==0.7.1
Pympler==1.0.1
PyNaCl==1.4.0
pyparsing==3.0.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.1
python-ffmpeg==1.0.11
python-multipart==0.0.5
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-image==0.19.3
scipy==1.9.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
starlette==0.17.1
streamlit==1.8.1
streamlit-webrtc==0.36.1
terminado==0.13.3
testpath==0.6.0
tifffile==2022.8.8
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typing-extensions==3.10.0.2
tzdata==2022.1
tzlocal==4.1
ujson==5.2.0
urllib3==1.26.9
uvicorn==0.17.6
validators==0.18.2
watchdog==2.1.7
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.1
widgetsnbextension==3.6.0
yarg==0.1.9
yarl==1.7.0
youtube-dl==2021.6.6
zipp==3.7.0
